+++ Lib/fontTools/svgLib/path/parser.py
