+++ Tests/svgLib/path/parser_test.py
