+++ Tests/svgLib/path/path_test.py
